Support skipping unmodified models during generation
Add skipUnmodifiedModels property to the <generate> Ant task.  Modify
BuildMps_GeneratorOptions to generate this property with default value
of "${mps.generator.skipUnmodifiedModels}", generate this property on
the project level with the default value of false.

This keeps the previous behavior (regenerating everything) if the user
does nothing, but allows overriding it by passing
-Dmps.generator.skipUnmodifiedModels=true to Ant.