MPS-28016 JUnitInProcessTermination_Test.test_terminate blinking
We do not want to provide the contract that #getCurrentRunner is not
already null.
The test itself is incorrect -- it seems that the data test is not started
execution when the check happens.

Answering the Artem comment which is removed -- yes, it is necessary to
run all the requests and to stop them consequtively in order to allow
the test server framework to handle the corresponding events correctly.
It is not in our place to ignore the requests at the client side,
expecting that the UI handle the absence of tests in a right way.