MPS-27810 consistency Assertion error in the xmlPersistence sample
junit sometimes sends several error notifications for the same tests.
e.g #onTestfailure, #onTestAssumptionFailure, etc.

note that xmlPersistence still fails though not with the ugly assertion
error. The reason is that now after artem platform improvements it does not get
initialized with environment and fails badly. But that is another
problem.