lang.refactoring: move runtime methods into RefactoringAccess
There's no reason to keep both RefactoringAccess and RefactoringUtil. Former, though an ugly singleton, is still better than latter as it is initialized with an access to MPS Platform,
therefore has a chance to access proper LanguageRegistry instance once we introduce an aspect for refactorings and stop loading their classed by name.