MPS-28295 VisibleModuleRegistry creates multiple Pattern on each check
VisibleModuleRegistry:
 * add creation of patterns at initialization
 * add empty module name check - name check does not make sense in this case and it can break Map because of duplicate values
 * update to java 8
 * code formatting

Pattern object is expensive to create for every check. Despite the fact, that there are not so many checks occurring, there are 37 extensions to jetbrains.mps.VisibleModuleMask and they do not change until application is closed, so recreating all of them on each check is not necessary.
So they can be just initialized and stored with component initialization.

VisibleModuleMask: annotation