Fixed MPS-28236, MPS-28252, MPS-28122
Changes in module cloning logic:
- Cloning for file based model root extracted in separate CopyFileBasedModelRootHelper class
- CopyNotSupportedException thrown only in if model root is in inappropriate state with meaningful
error message, that then will be shown to user
- Replaced "something wrong happened" error dialog with meaningful popup
- Reimplement lost logic of copying files in java_source_stub & java_classes model roots
- Reference adjusting works well for read-only models - checks for read-only flags had no sense
- Make facets copying via memento - however it may not work for complicated facet, current facets are
so trivial that no need in custom logic of it cloning. One exception is `custom generation` facet that allows
to specify genplan with model reference. To support adjusting of this reference we need to make
ReferenceUpdater extensible.