StaticReference: introduce fair repository handling
When a node is detached and references are replaced with 'direct' nodes, don't use context-agnostic getTargetSModel() as it works with a global repository (ref.resolve(null)) which breaks the story of detached model being modified without need for model access (source node in a detached model, target node comes from an attached model, and model.getNode fails with IMAE)

There are 2 subclasses for SReferenceBase, with DynamicReference being (almost) independent of mature/immature handling. It's odd to override makeDirect() in DynamicReference to make it no-op, when the only implementation of the method makes sense for StaticReference subclass. Moved methods appropriately.

Pending: is there any reason for DynamicReference to subclass SReferenceBase (or access immatureTargetNode, for the matter)?