RepositoryStateCacheUtils: drop last use; nag clients to abandon its use
ClassifierScopeUtils is used a lot, but the approach to cache inside a static utilty class is plain wrong (let alone static utility class itself is poor programming practice).
Instead of caching, callers shall be written in an effective way (e.g. building a set of supeclasses once and reusing it).