Getting rid of the custom packaging idea plugins in the build language.
It is totally the layout information, thus to be transferred to the
layout. Moreover due to the complexity of the per-module custom
packaging I do not think that it makes sense to keep this logic at all.
It is going to be either all modules are packed by MPS or all modules
are packed by the user.