MPS-28602 Scope cache aware of a model read
As long as there could be no change in a model during read, we are safe to cache objects like scope and resolved nodes.
Since Scope instances were never intended for multi-thread use, keep a thread-local copy.
It's safe to resort to earlier implementation in case anything is not in the state we expect (e.g. reference's model != null but is not in a repo).