TextGen: keep record of generated files in 'generated' cache
Instead of walking FS to collect generated files, record them in the file.
With that, we get fine control over what is stale and have to be removed.
Essential part of arbitrary output locations per text unit functionality (towards BL packages not tied to a model) - if it's textgen to tell output folder, we have no chance to figure out stale files for 'unchanged' models than, and can not treat all files under e.g. source_gen as stale.