Make: abstraction for delta key
Would like to stop using file path as key as it's not convenient for scenarios when model output is not bound to single folder - no clear idea which folder to use as delta key then.
Use of source_gen top-level folder imposes need to expose it from GenerationTargetFacet which I'd like to avoid.