FS listener: drop dependency listener
No reason to prioritize listener for backup file, it merely schedules update of a primary DS. Without priority, either notification comes to primary listener first, and then update does the job as usual, and event.notify() for 'notified' listener is no-op. If notification for the backup file comes first, then it schedules primary listener update and works as before.