Language module no longer owner for Generators originating from its MD
MPSModuleOwner concept is a marker with uncertain contract, the fact Language module (and only Language) could be an owner for other modules is an implementation detail everybody has to be aware of. Now, it's plain and straightforward, modules and their owners indicating origin/provider of the modules, with no nesting.
Eventually we won't need MPSModuleOwner concept, which looks like a mechanism to support singleton repository and modules coming from different locations (i.e. deployed and project). Once we can have separate repositories for modules with same identity, we can throw MPSModuleOwner away.