Provisional code to avoid re-save modules on load
When a module file is reported as changed on startup (e.g. due to checkout of different project version), SLibrary/ProjectModuleFileChangeListener react and push a new MD instance to a loaded module. This is not bad per se (as it helped to reveal few outdated modules), but idea of 'reloadFromDisk' is that module considered clean once it's done, not 'changed.'