Fix MPS-28961 Idea plugins: library "mps.jetbrains.mps.baseLanguage.references.runtime_auto" is constantly added in iml when I write on base language
In one of previous commits the affected code (accidentally?) switched
from comparing paths to comparing IFile objects by contains(equals).
The problem was that one IFile was IdeaFile and the other was JarEntryFile
for the same path. The latter happened because for deployed modules
java.io (plain, non-caching) filesystem is used.