Show In Project: ensure there's a tree node for a module we navigate to.
It's odd to keep and update global file-module map just for the sake of single 'show in logical view' action. Besides, the fact we know file is mapped to a module doesn't help to ensure the module is present in the logical view. Now, we walk the project tree to find if there's a node that matches a module file.