MPS-29051 Exception while creating substitute actions in class jetbrains.mps.nodeEditor.cellMenu.SEmptyContainmentSubstituteInfo
Since the MPS-28867 is open and rebuild language does not recompile java files that did not change, we return the original constructor so users won't face the exceptions.
The generated code will use the new constructor though so we could remove the old constructor again in the future an not face this problem again.

There is workaround for the issue which is "Recompile Classes" but it is hard to find and there is no discovarable guide for the users.