module-node projections: do not expose empty models
This is a workaround to avoid referencing aspect classes that were not generated.
Descriptor references aspects based solely on presence of an aspect model. However, aspect class is generated only iff there are true uses of an aspect nodes.
There's a hack to add main aspect language to any aspect model just to make them generated, but now it's gone.