Force generation of editor aspect class
Even though there's no editor-related stuff in the editor aspect model, force generation of the aspect class as generation of Language class decides to reference it based solely on existence of non-empty editor model.
Indeed, it's odd to keep some test classes in the editor aspect, but it's too much for me now to refactor these tests.