MPS-29461 no need to show CopyOutcome annotation for an attribute
Makeup facet recognizes CopyOutcome for an attributed initial node of a textgen. Though technically it's ok to have an attribute as a textgen starter, it never happens (that's not what attributes are designed for). To avoid confusion, just hide CopyOutcome annotation from locations that are likely erroneous (incliding CopyOutcome annotation itself, notice excessive'Add @CopyOutcome annotation' action in schreenshots from the issue.