break [generator]-[kernel] dependency cycle
Generator needs module implementations, module implementations need descriptors, descriptors, namely GeneratorDescriptor, needs generator api and impl classes (priority rules). As long as these are part of module description, moved into kernel, though better way would be another (kernel2, with generator) module or Generator as a part of [generator-engine] (which is not possible at the moment as there are a lot of uses of the class from the [kernel]).