Ant-MPS startup code: get ready to move out of MPS-aware module
The code to collect MPS libraries and to start an MPS shall not demand MPS in its classpath.
Now AntBootstrap uses classes of j.m.core.tool.environment module with classpath dependencies to kernel, platform, etc.
Though I don't see a reason to keep UrlClassLoader and not to use java.net.URLClassLoader, left it for a future excercise, yet simplified some bits not to drag along a lot of unrelated code.
With no evidence of performance gain, and whatever bugs left in this copy of IDEA classes (now they are completely different), I don't know if we shall stick to this home-grown alternative to standard classes.