Ant-MPS startup code: keep its use of MPS classpath
ant-mps-common doesn't need MPS kernel in classpath. Its purpose is to build an MPS classpath to get MPS running, and it's odd to build an MPS classpath just to fire a code that uses MPS to built its classpath once again. Code to build a classpath shall have minimalistic dependencies to the rest of MPS.