Packaged module test: expose classes of stub model root right.
Deployment classpath is no longer exposed as java stub root (AM used to expose all additionalJavaStubPaths). If one references exact jar in a stub model root, there's a need to make sure this jar is present in the build and is properly referenced. Though 'properly' is a bit vague here, in no case we shall expose whole dir neither as classpath nor as stub, and 'properly' in this scenario means 'point to exact stub by its name', so that ModuleMiner code could match deloyed jars with model root entry.